refactor: 각 레이어 역할에 맞게 패키지 이동
