+++ app/src/main/AndroidManifest.xml
@@ -18,11 +18,11 @@
         tools:targetApi="31">
 
         <receiver
-            android:name=".ui.view.receiver.AlarmReceiver"
+            android:name=".ui.receiver.AlarmReceiver"
             android:exported="true"/>
 
         <receiver
-            android:name=".ui.view.receiver.BootReceiver"
+            android:name=".ui.receiver.BootReceiver"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
