+++ app/src/main/java/woowacourse/movie/domain/datasource/TicketDataSource.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.ui.view.data
+package woowacourse.movie.domain.datasource
 
 import woowacourse.movie.domain.ticket.Ticket
 
-interface TicketDataAdapter {
+interface TicketDataSource {
     fun insert(ticket: Ticket): Long
 
     fun getTicket(id: Long): Ticket
