+++ app/src/main/java/woowacourse/movie/ui/util/ErrorMessage.kt
@@ -1,10 +1,10 @@
-package woowacourse.movie.ui.view.util
+package woowacourse.movie.ui.util
 
 @JvmInline
 value class ErrorMessage(
     private val cause: String,
 ) {
-    fun notProvided() = "$cause ${ERROR_MESSAGE_NO_DATA_FORMAT}"
+    fun notProvided() = "$cause $ERROR_MESSAGE_NO_DATA_FORMAT"
 
     fun notSelected() = "$cause ${ERROR_MESSAGE_NOT_SELECTED_YET_FORMAT}T"
