+++ app/src/main/java/woowacourse/movie/ui/view/history/ReservationHistoryPresenter.kt
@@ -1,13 +1,13 @@
 package woowacourse.movie.ui.view.history
 
-import woowacourse.movie.ui.view.data.TicketDataAdapter
+import woowacourse.movie.domain.datasource.TicketDataSource
 import kotlin.concurrent.thread
 
 class ReservationHistoryPresenter(
     private val view: ReservationHistoryContract.View,
-    private val ticketDataAdapter: TicketDataAdapter,
+    private val ticketDataSource: TicketDataSource,
 ) : ReservationHistoryContract.Presenter {
     override fun presentScreen() {
-        thread { view.updateScreen(ticketDataAdapter.getAll()) }
+        thread { view.updateScreen(ticketDataSource.getAll()) }
     }
 }
